Reorganize writers
